+++ CVNet_Trainer/cmdlineopt.h
@@ -5,18 +5,18 @@ namespace CmdLineOpt {
 	void CmdLineOpt(int argc, const char* argv[]);
 
 	extern float drop_rate;
-	extern uint32_t growth_rate;
-	extern std::vector<int> params;
+	extern size_t growth_rate;
+	extern std::vector<size_t> params;
 	
 	enum TYPE { DENSENET, OTRANET, PIRAMIDAL };
-	extern int type_net;
+	extern size_t type_net;
 
-	extern int  epoch;
-	extern int  batch_size;
+	extern size_t  epoch;
+	extern size_t  batch_size;
 	extern bool cpu;
 	extern bool verbose;
-	extern int batch_norm;
-	extern int overwrite;
+	extern size_t batch_norm;
+	extern size_t overwrite;
 	extern float percent_to_train;
 	extern float learning_rate;
 	extern std::string dataset_path;
